+++ extensions/paste-as-plain-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Paste as Plain text Changelog
 
+## [Function Optimisation] - 2024-06-06
+
+-  Update Paste as Number: Now all the numbers will be extracted and stitched together for display
+
 ## [Fix Hotkey Error] - 2024-05-27
 
 -  Fix the problem of using hotkeys to invoke commands